+++ src/zope/testrunner/formatter.py
@@ -766,7 +766,7 @@ class FakeTest:
 # dependency on subunit.
 try:
     import subunit
-    from subunit.iso8601 import Utc
+    from iso8601 import UTC
     subunit.StreamResultToBytes
 except (ImportError, AttributeError):
     subunit = None
@@ -880,7 +880,7 @@ class SubunitOutputFormatter:
 
     def __init__(self, options, stream=None):
         if subunit is None:
-            raise Exception('Requires subunit 0.0.11 or better')
+            raise Exception('Requires python-subunit 1.4.3 or better')
         if testtools is None:
             raise Exception('Requires testtools 0.9.30 or better')
         self.options = options
@@ -893,7 +893,7 @@ class SubunitOutputFormatter:
         # Used to track the last layer that was set up or torn down. Either
         # None or (layer_name, last_touched_time).
         self._last_layer = None
-        self.UTC = Utc()
+        self.UTC = UTC
         # Content types used in the output.
         self.TRACEBACK_CONTENT_TYPE = ContentType(
             'text', 'x-traceback', {'language': 'python', 'charset': 'utf8'})